+++ src/Model/Model.cpp
@@ -890,7 +890,7 @@ bool Model::_serializeAscii(std::ostream& os, bool /*verbose*/) const
 
     if (!cova->getFlagAniso()) continue;
 
-    for (unsigned int idim = 0; ret && idim < getNDim(); idim++)
+    for (size_t idim = 0; ret && idim < getNDim(); idim++)
       ret = ret && _recordWrite<double>(os, "", cova->getAnisoCoeff(idim));
     ret = ret && _commentWrite(os, "Anisotropy Coefficients");
     ret = ret && _recordWrite<int>(os, "Anisotropy Rotation Flag", (int)cova->getFlagRotation());
@@ -898,8 +898,8 @@ bool Model::_serializeAscii(std::ostream& os, bool /*verbose*/) const
     if (!cova->getFlagRotation()) continue;
 
     // Storing the rotation matrix by Column (compatibility)
-    for (unsigned int idim = 0; ret && idim < getNDim(); idim++)
-      for (unsigned int jdim = 0; ret && jdim < getNDim(); jdim++)
+    for (size_t idim = 0; ret && idim < getNDim(); idim++)
+      for (size_t jdim = 0; ret && jdim < getNDim(); jdim++)
         ret = ret && _recordWrite<double>(os, "", cova->getAnisoRotMatElement(jdim, idim));
     ret = ret && _commentWrite(os, "Anisotropy Rotation Matrix");
   }
